+++ MODEL/SQuery.cs
@@ -3,12 +3,13 @@ public class SQuery
 {
     public string CodigoProducto { get; set; }
     public string Descripcion { get; set; }
-    public double PrecioFarmaciaCat { get; set; }
+    public decimal PrecioFarmaciaCat { get; set; }
     public int CodigoSucursal { get; set; }
     public int Existencia { get; set; }
-    public double PrecioFarmacia { get; set; }
-    public double Margen { get; set; }
+    public decimal PrecioFarmacia { get; set; }
+    public decimal Margen { get; set; }
     public string CodigoRelacionado { get; set; }
     public int CantidadPastillero { get; set; }
-    public double PrecioActual { get; set; }
+    public decimal PrecioActual { get; set; }
+
 }
